requirements/base.txt
jinja2==2.8
tqdm==3.7.1
pyyaml==3.11
ghp-import==0.4.1
sd-python-wrapper==0.1.33
watchdog
